+++ simpful/simpful.py
@@ -1,8 +1,8 @@
-from numpy import array, argmin, argmax, linspace, exp
+from .fuzzy_sets import FuzzySet, MF_object, Sigmoid_MF, InvSigmoid_MF, Gaussian_MF, InvGaussian_MF, DoubleGaussian_MF
+from .rule_parsing import curparse, preparse, postparse
+from numpy import array, linspace
 from scipy.interpolate import interp1d
 from collections import defaultdict
-import re
-import numpy as np
 try:
 	import seaborn as sns
 except ImportError:
@@ -16,79 +16,6 @@ class UndefinedUniverseOfDiscourseError(Exception):
 
 	def __init__(self, message):
 		self.message = message
-
-
-class MF_object(object):
-
-	def __init__(self):
-		pass
-
-	def __call__(self, x):
-		ret = self._execute(x)
-		return min(1, max(0, ret))
-		
-
-###############################
-# USEFUL PRE-BAKED FUZZY SETS #
-###############################
-
-def gaussian(x, mu, sig):
-    return np.exp(-np.power(x - mu, 2.) / (2 * np.power(sig, 2.)))
-
-class Sigmoid_MF(MF_object):
-
-	def __init__(self, c=0, a=1):
-		self._c = c
-		self._a = a
-		
-	def _execute(self, x):
-		return 1.0/(1.0 + exp(-self._a*(x-self._c))) 
-
-class InvSigmoid_MF(MF_object):
-
-	def __init__(self, c=0, a=1):
-		self._c = c
-		self._a = a
-		
-	def _execute(self, x):
-		return 1.0 - 1.0/(1.0 + exp(-self._a*(x-self._c))) 
-
-class Gaussian_MF(MF_object):
-
-	def __init__(self, mu, sigma):
-		self._mu = mu
-		self._sigma = sigma
-
-	def _execute(self, x):
-		return gaussian(x, self._mu, self._sigma)
-
-class InvGaussian_MF(MF_object):
-
-	def __init__(self, mu, sigma):
-		self._mu = mu
-		self._sigma = sigma
-
-	def _execute(self, x):
-		return 1.-gaussian(x, self._mu, self._sigma)
-
-class DoubleGaussian_MF(MF_object):
-
-	def __init__(self, mu1, sigma1, mu2, sigma2):
-		self._mu1 = mu1
-		self._sigma1 = sigma1
-		self._mu2 = mu2
-		self._sigma2 = sigma2
-
-	def _execute(self, x):
-		first = gaussian(x, self._mu1, self._sigma1)
-		second = gaussian(x, self._mu2, self._sigma2)
-		return first*second
-
-
-###############################
-# USEFUL PRE-BAKED FUZZY SETS #
-###############################
-
 
 
 class LinguisticVariable(object):
@@ -168,78 +95,6 @@ class LinguisticVariable(object):
 
 	def __repr__(self):
 		return "L.V.: "+self._concept
-
-
-
-class FuzzySet(object):
-
-	def __init__(self, points=None, function=None, term="", high_quality_interpolate=True, verbose=False):
-
-		self._term = term
-
-		if points is None and function is not None:
-			self._type = "function"
-			self._funpointer = function
-			#self._funargs	= function['args']
-			return
-
-
-		if len(points)<2: 
-			print("ERROR: more than one point required")
-			exit(-1)
-		if term=="":
-			print("ERROR: please specify a linguistic term")
-			exit(-3)
-		self._type = "pointbased"
-		self._high_quality_interpolate = high_quality_interpolate
-
-		"""
-		if verbose:
-			if len(points)==1: # singleton
-				pass
-			elif len(points)==2: # triangle
-				print(" * Triangle fuzzy set required for term '%s':" % term, points)
-				self._type = "TRIANGLE"
-			elif len(points)==3: # trapezoid
-				print(" * Trapezoid fuzzy set required for term '%s':" % term, points)
-				self._type = "TRAPEZOID"
-			else:
-				print(" * Polygon set required for term '%s':" % term, points)
-				self._type = "POLYGON"
-		"""
-
-		self._points = array(points)
-		
-
-	def get_value(self, v):
-
-		if self._type == "function":
-			return self._funpointer(v)
-
-		if self._high_quality_interpolate:
-			return self.get_value_slow(v)
-		else:
-			return self.get_value_fast(v)
-
-	def get_value_slow(self, v):		
-		f = interp1d(self._points.T[0], self._points.T[1], 
-			bounds_error=False, fill_value=(self._points.T[1][0], self._points.T[1][-1]))
-		result = f(v)
-		return(result)
-
-	def get_value_fast(self, v):
-		x = self._points.T[0]
-		y = self._points.T[1]
-		N = len(x)
-		if v<x[0]: return self._points.T[1][0] 
-		for i in range(N-1):
-			if (x[i]<= v) and (v <= x[i+1]):
-				return self._fast_interpolate(x[i], y[i], x[i+1], y[i+1], v)
-		return self._points.T[1][-1] # fallback for values outside the Universe of the discourse
-
-	def _fast_interpolate(self, x0, y0, x1, y1, x):
-		#print(x0, y0, x1, y1, x); exit()
-		return y0 + (x-x0) * ((y1-y0)/(x1-x0))
 
 
 class FuzzySystem(object):
@@ -403,159 +258,6 @@ class FuzzySystem(object):
 
 		fig.tight_layout()
 		fig.savefig(outputfile)
-
-
-class Clause(object):
-
-	def __init__(self, variable, term, verbose=False):
-		self._variable = variable
-		self._term = term
-
-	def evaluate(self, FuzzySystem, verbose=False, operators=None):
-		try:
-			ans = FuzzySystem._lvs[self._variable].get_values(FuzzySystem._variables[self._variable])
-		except KeyError:
-			raise Exception("ERROR: variable '" + self._variable + "' not defined.\n"
-				+ " ---- PROBLEMATIC CLAUSE:\n"
-				+ str(self))
-		if verbose: 
-			print("Checking if", self._variable,)
-			print("whose value is", FuzzySystem._variables[self._variable],)
-			print("is actually", self._term)
-			print("answer:", ans[self._term])
-		try:
-			return ans[self._term]
-		except KeyError:
-			raise Exception("ERROR: term '" + self._term + "'' not defined.\n"
-				+ " ---- PROBLEMATIC CLAUSE:\n"
-				+ str(self))
-
-	def __repr__(self):
-		return "c.(%s IS %s)" % (self._variable, self._term)
-
-
-class Functional(object):
-
-	def __init__(self, fun, A, B, operators=None):
-		self._A = A
-		self._B = B
-
-		if operators is None:
-			self._fun = fun
-		else:
-			if "AND_PRODUCT" in operators: 
-				if fun=="AND":
-					self._fun = "AND_p"
-				else:
-					self._fun = fun
-			else:
-				self._fun = fun
-
-	def evaluate(self, FuzzySystem):
-		if self._A=="":
-			# support for unary operators
-			# print("Unary detected")
-			B = self._B.evaluate(FuzzySystem)
-			return array(eval(self._fun+"(%s)" % B))
-		else:
-			A = self._A.evaluate(FuzzySystem)
-			B = self._B.evaluate(FuzzySystem)
-			return array(eval(self._fun+"(%s, %s)" % (A,B)))
-		
-	def __repr__(self):
-		return "f.(" + str(self._A) + " " + self._fun + " " + str(self._B) + ")"
-
-
-# basic definitions o f 
-def OR(x,y): return max(x, y)
-def AND(x,y): return min(x, y)
-def AND_p(x,y): return x*y
-def NOT(x): return 1.-x
-
-
-def preparse(STRINGA):
-	# extract the antecedent
-	return STRINGA[STRINGA.find("IF")+2:STRINGA.find("THEN")].strip()
-
-def postparse(STRINGA, verbose=False):
-	# extract the consequent
-	stripped = STRINGA[STRINGA.find("THEN")+4:].strip("() ")
-	if STRINGA.find("THEN") == -1:
-		raise Exception("ERROR: badly formatted rule, please check capitalization and syntax.\n"
-					+ " ---- PROBLEMATIC RULE:\n"
-					+ STRINGA)
-	return stripped[:stripped.find("IS")].strip(), stripped[stripped.find("IS")+2:].strip()
-
-def find_index_operator(string, verbose=False):
-	if verbose: print(" * Looking for an operator in", string)
-	pos = 0
-	par = 1
-	while(par>0):
-		pos+=1
-		# if pos>=len(string):
-			#print(pos, pos2)
-			# raise Exception("badly formatted rule, terminating")
-		if string[pos]==")": par-=1
-		if string[pos]=="(": par+=1
-	pos2 = pos
-	while(string[pos2]!="("):
-		pos2+=1
-	return pos+1, pos2
-
-
-def curparse(STRINGA, verbose=False, operators=None):
-
-	# base case
-	if STRINGA=="": return "" 
-
-	STRINGA = STRINGA.strip()
-	if STRINGA[0]!="(": STRINGA="("+STRINGA
-	if STRINGA[-1]!=")": STRINGA=STRINGA+")"
-	
-	#regex = re.compile("^\([a-z,_,A-Z]* IS [a-z,_,A-Z]*\)$")
-	regex = re.compile("^\([a-z,_,A-Z,0-9]*\s*IS\s*[a-z,_,A-Z,0-9]*\)$")
-	if regex.match(STRINGA):
-		
-		if verbose:	print(" * Regular expression is matching with single atomic clause:", STRINGA)
-
-		# base case
-		variable = STRINGA[1:STRINGA.find("IS")].strip()
-		term	 = STRINGA[STRINGA.find("IS")+3:-1].strip()
-		ret_clause = Clause(variable, term, verbose=verbose)
-		if verbose:	print(" * Rule:", ret_clause)
-		return ret_clause
-
-	else:
-
-		# there can be two explanations: missing parentheses, or sub-expression
-		if verbose:	print(" * Regular expression is not matching with single atomic clause:", STRINGA)
-		
-		# try recursion
-		removed_parentheses = STRINGA[STRINGA.find("(")+1:STRINGA.rfind(")")].strip()
-		
-		# if (removed_parentheses.find("(")==-1): return curparse("("+removed_parentheses+")")
-
-		if verbose: print("  - After parentheses removal:", removed_parentheses)
-
-		if removed_parentheses[:3]=="NOT":
-			beginindop = 0
-			endindop = 3
-			#if verbose: print( " * Detected unary operator NOT")
-		else:
-			try:
-				beginindop, endindop = find_index_operator(removed_parentheses, verbose=verbose)
-			except IndexError:
-				raise Exception("ERROR: badly formatted rule, please check capitalization and syntax.\n"
-					+ " ---- PROBLEMATIC RULE:\n"
-					+ STRINGA)
-
-		firsthalf = removed_parentheses[:beginindop].strip()
-		secondhalf = removed_parentheses[endindop:].strip()
-		operator = removed_parentheses[beginindop:endindop].strip()
-		if verbose:	print("  -- Found %s *%s* %s" % (firsthalf, operator, secondhalf))
-		
-		return Functional(operator, curparse(firsthalf, verbose=verbose, operators=operators), curparse(secondhalf, verbose=verbose, operators=operators), operators=operators)
-
 
 if __name__ == '__main__':
